Move to package layout for mypy compatibility
