+++ .gitignore
@@ -55,3 +55,5 @@ coverage.xml
 # Sphinx documentation
 docs/_build/
 
+# macOS files
+.DS_Store
