--- MANIFEST
@@ -1,7 +1,0 @@
-# file GENERATED by distutils, do NOT edit
-CHANGES.md
-LICENSE
-README.md
-setup.py
-src/inject.py
-src/inject.pyi
