--- MANIFEST.in
@@ -1,3 +1,0 @@
-include LICENSE
-include CHANGES.md
-include README.md
